add subrun summary information to the output file
Add a subrun tree with total/good POT information to the output tree.
This implementation is fragile, and accesses the TFile from the
gallery::Event, pending an implementation of proper subrun access in
gallery. Thanks to N. Tagg for suggesting this approach!

Since we may be processing multiple files, we inspect each event's
associated files for the subrun list and build up a cache of
run/subrun tuples. This may not be valid for all MC if those tuples
are not unique, so the caching strategy may need adjustment.