import sbnanalysis framework to run in a larsoft analyzer
Import the sbnanalysis gallery framework into LArSoft, so that the event loop runs
inside an analyzer module. The Processor framework structure is unchanged, except
trivially using art::Events rather than gallery::Events.

To do:
* Use the TFileService for the output file
* Reimplement the subrun data using the proper LArSoft accessors

Note that the ProviderManager is no longer needed, as services are now available
in the usual way, with an art::ServiceHandle.