Major changes towards multiflash matching support
Currently module is partially broken, some more changes are needed to
bring it back to previous working status.

Idea now is to get all the flashes at once and store them on the
vector of simpleFlashes. Since SBND volumes are optically isolated,
three sets of flashes are produced with ophits from: both volumes,
left volume and right volume.

Not yet working, but computation of metrics will be stored for each
event, and reused if possible.