Use GeoAlgo library in larcorealg
The GeoAlgo library is also in larcorealg. Previously, we were using the one in sbncode/OpT0Finder/flashmatch, although this creates problems when running with other code that calls classes in larcorealg/GeoAlgo. We should change the namespace of the one in sbncode/OpT0Finder/flashmatch, but for now I am simply switching to the larcorealg one.