The Metrics now come in 3D!!!!
Create a 3D voxelised map per metric, that takes into account the
X,Y,Z coordinates of a slice of charge to score the association with a
flash.

Create a new class in the template generator to simplify the code and
aid to implement this new feature.

Quite a big commit. It's fully functional, however not yet enabled as
it requires quite a bit of testing and debbuging.