Update SBND BNB info fcl to use a new process name
For the Spring 2025 production patch release, the SBND BNB info fcl needs to be re-run to pick up the BNB FOM calculation. To enable this, we use a new process name in that fcl. We also drop the old version of the BNB info data products to reduce the possibility that they get picked up by mistake.